requirements.txt
streamlit
plotly
pandas
google-auth
gspread